# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  login: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  quantity: number;
  category: string;
  subcategory: string;
  description: string;
  imageUrl: string;
}

export interface ProductInput {
  /** @minLength 2 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  quantity: number;
  /** @minLength 1 */
  category: string;
  subcategory: string;
  /** @minLength 1 */
  imageUrl: string;
}

export interface ProductUpdate {
  /** @minLength 2 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  quantity?: number;
  /** @minLength 1 */
  category?: string;
  subcategory?: string;
  /** @minLength 1 */
  imageUrl?: string;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  parentId: number | null;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  parentId?: number | null;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  parentId?: number | null;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  lineTotal: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  new: 'new',
  picking: 'picking',
  delivery: 'delivery',
  completed: 'completed',
} as const;

export interface Order {
  id: number;
  clientName: string;
  phone: string;
  address: string;
  totalSum: number;
  status: OrderStatus;
  createdAt: string;
  items: OrderItem[];
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  /** @minLength 2 */
  clientName: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 5 */
  address: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface OrderCreated {
  orderId: number;
  totalSum: number;
}

export type OrderStatusInputStatus = typeof OrderStatusInputStatus[keyof typeof OrderStatusInputStatus];


export const OrderStatusInputStatus = {
  new: 'new',
  picking: 'picking',
  delivery: 'delivery',
  completed: 'completed',
} as const;

export interface OrderStatusInput {
  status: OrderStatusInputStatus;
}

export interface OrderSummary {
  totalOrders: number;
  totalRevenue: number;
  newOrders: number;
  pickingOrders: number;
  deliveryOrders: number;
  completedOrders: number;
}

export interface ErrorResponse {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListProductsParams = {
search?: string;
category?: string;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  new: 'new',
  picking: 'picking',
  delivery: 'delivery',
  completed: 'completed',
} as const;