# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: List products
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a catalog product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a catalog product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a catalog product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Product is referenced by an order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories:
    get:
      operationId: listCategories
      tags: [catalog]
      summary: List product categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create a category or subcategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [admin]
      summary: Update a category or subcategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCategory
      tags: [admin]
      summary: Delete a category or subcategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Category deleted
        "400":
          description: Category is still used by products or subcategories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for managers
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [new, picking, delivery, completed]
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a customer order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreated"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        price:
          type: number
        quantity:
          type: number
        category:
          type: string
        subcategory:
          type: string
        description:
          type: string
        imageUrl:
          type: string
      required: [id, name, price, quantity, category, subcategory, description, imageUrl]
    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        price:
          type: number
          minimum: 0
        quantity:
          type: number
          minimum: 0
        category:
          type: string
          minLength: 1
        subcategory:
          type: string
        imageUrl:
          type: string
          minLength: 1
      required: [name, price, quantity, category, subcategory, imageUrl]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        price:
          type: number
          minimum: 0
        quantity:
          type: number
          minimum: 0
        category:
          type: string
          minLength: 1
        subcategory:
          type: string
        imageUrl:
          type: string
          minLength: 1
    Category:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        parentId:
          type: ["number", "null"]
      required: [id, name, parentId]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        parentId:
          type: ["number", "null"]
      required: [name]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        parentId:
          type: ["number", "null"]
    OrderItem:
      type: object
      properties:
        productId:
          type: number
        productName:
          type: string
        quantity:
          type: number
        price:
          type: number
        lineTotal:
          type: number
      required: [productId, productName, quantity, price, lineTotal]
    Order:
      type: object
      properties:
        id:
          type: number
        clientName:
          type: string
        phone:
          type: string
        address:
          type: string
        totalSum:
          type: number
        status:
          type: string
          enum: [new, picking, delivery, completed]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [id, clientName, phone, address, totalSum, status, createdAt, items]
    OrderInput:
      type: object
      properties:
        clientName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 5
        address:
          type: string
          minLength: 5
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required: [clientName, phone, address, items]
    OrderItemInput:
      type: object
      properties:
        productId:
          type: number
        quantity:
          type: number
          minimum: 1
      required: [productId, quantity]
    OrderCreated:
      type: object
      properties:
        orderId:
          type: number
        totalSum:
          type: number
      required: [orderId, totalSum]
    OrderStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [new, picking, delivery, completed]
      required: [status]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...