# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 2 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  quantity?: number;
  /** @minLength 1 */
  category?: string;
  subcategory?: string;
  /** @minLength 1 */
  imageUrl?: string;
}